I don't like this cast, but does not compile otherwise on g++ 4.7.3.

+++ toulbar2/src/tb2arithmetic.hpp
@@ -41,7 +41,7 @@ public:
     bool verify();
     
     double  computeTightness() {
-    	tight = (double) penalty * abs(permitted.size() - x->getDomainSize()) / x->getDomainSize();
+    	tight = (double) penalty * abs((long int)permitted.size() - x->getDomainSize()) / x->getDomainSize();
     	return tight;
     }
